+++ src/general.hpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/09 17:35:25 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/04/15 17:50:42 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/04/15 21:15:00 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,6 +46,8 @@
 #define TXT_BOLD_CYAN "\033[1;36m"
 //*		~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+#define FILETYPES "../conf/mime.types"
+
 enum e_methods
 {
 	GET,
@@ -60,5 +62,6 @@ struct s_request_info
 	std::string http_version;
 };
 
+std::vector <std::string> cpp_split(std::string full_line, char splitter);
 
 #endif