+++ src/responses/responses.hpp
@@ -1,2 +1,2 @@
 #include "../general.hpp"
-
+#include "../Filetypes.hpp"