+++ src/utils.cpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/06 20:30:58 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/04/11 19:07:06 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/04/15 21:18:54 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,14 @@ std::vector <std::string> cpp_split(std::string full_line, char splitter)
 	std::string::iterator begin = full_line.begin();
 	std::string::iterator end = full_line.begin();
 	
-
+	while (end != full_line.end() && *end == splitter)
+	{
+		end++;
+		begin++;
+	}
 	while (end != full_line.end())
 	{
+
 		while (end != full_line.end() && *end != splitter)
 			end++;
 		if (end != full_line.end())